// FE-navigation3/src/app/api/models/sample-image-diagnosis-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PdfFile } from '../models/pdf-file';
export interface SampleImageDiagnosisDto {
  bodySystemNames: Array<string>;
  caustryFile?: PdfFile;
  diagnosis: string;
  hasAnnotation: boolean;
  id: number;
  isFavorite: boolean;
  keyWords?: string | null;
  name: string;
  note?: string | null;
  organName: string;
}